requirements.txt
alembic==1.7.3
aniso8601==9.0.1
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
blinker==1.4
cachetools==5.2.0
certifi==2022.6.15
cffi==1.14.6
charset_normalizer==2.1.1
click==8.0.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.0
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask==2.0.1
fsspec==2022.5.0
gcsfs==2022.5.0
google_auth==2.11.0
google==3.0.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==4.0.1
kubernetes==24.2.0
Mako==1.1.5
MarkupSafe==2.0.1
marshmallow-sqlalchemy==0.26.1
marshmallow==3.13.0
oauthlib==3.2.0
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyasn1_modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.7
pyrsistent==0.18.0
python_dateutil==2.8.2
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0
requests_oauthlib==1.3.1
requests==2.28.1
rsa==4.9
setuptools==65.3.0
six==1.16.0
SQLAlchemy==1.4.25
tomli==2.0.0
urllib3==1.26.12
websocket_client==1.4.0
Werkzeug==2.0.1
